requirements.txt
django>=4.2,<6.1
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
httpx>=0.27.0
